+++ communication/src/one/microstream/com/ComProtocolProviderCreator.java
@@ -13,7 +13,7 @@ public interface ComProtocolProviderCreator<C>
 		String                                name                  ,
 		String                                version               ,
 		ByteOrder                             byteOrder             ,
-		long                                  inactivityTimeout     ,
+		int                                   inactivityTimeout     ,
 		PersistenceIdStrategy                 idStrategy            ,
 		PersistenceTypeDictionaryViewProvider typeDictionaryProvider,
 		ComProtocolCreator                    protocolCreator
@@ -47,7 +47,7 @@ public interface ComProtocolProviderCreator<C>
 			final String                                name                  ,
 			final String                                version               ,
 			final ByteOrder                             byteOrder             ,
-			final long                                  inactivityTimeout     ,
+			final int                                   inactivityTimeout     ,
 			final PersistenceIdStrategy                 idStrategy            ,
 			final PersistenceTypeDictionaryViewProvider typeDictionaryProvider,
 			final ComProtocolCreator                    protocolCreator
